requirements.txt
pandas
numpy
scikit-learn
matplotlib
flask
xgboost
weasyprint==66.0
gunicorn
google-cloud-storage